pyproject.toml
[project]
name = "partages"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.0"
dependencies = [
    "pyyaml",
    "datasets",
]
